Ignore files modified by test suite
These are some JSON files that seem to be modified after each test
run.  I suspect that they should not be modified at all when running
tests (if they are input files, they should be left alone; if they are
output files, they shouldn't be in VCS), but PCE is doing just that.
Need further investigation.