Recommend sourcing .env when running docker compose up
The documented format for docker env files is `KEY=VAL`, according to
docs:

https://docs.docker.com/compose/environment-variables/env-file/

But we need `export KEY=VAL` to support testing with tox, and to avoid
duplication.

It also seems that docker compose does not mind the `export KEY=VAL`
format in .env files after all. Perhaps it is an undocumented feature
that will go away, so recommend sourcing anyway.