Move create_app() to module top-level
This is a work-around with testing in mind. We do not really use the
message queue, at least right now, in our little unit/integration test
suite. We do use the database.  With this restructuring, we do not
have to bring up the message queue for testing, just the DB.

When we are ready to test things with the message queue, we might want
to turn this back though.