Use `expose` instead of `ports` for tox-docker
tox-docker 5.0.0 has deprecated `ports`; we need to use `expose`
instead.  This is a good change; with `expose` the services we use
when testing should not conflict with any pre-existing services.

We could have pinned to an older version of tox-docker, but older
tox-docker sometimes errors out if the right version of
requests (<=2.31.0) is not available.

https://github.com/docker/docker-py/issues/3256

A fallout of this change is that we can't use MONGODB_CONNSTRING
anymore, since forming a URI using environment variables from tox
appears to be rather awkward.

https://github.com/tox-dev/tox-docker/issues/187